+++ src/docs/asciidoc/api/CampaignCategory-API.adoc
