+++ src/docs/asciidoc/api/CampaignType-API.adoc
