+++ src/docs/asciidoc/index.adoc
@@ -7,25 +7,20 @@
 :sectlinks:
 :docinfo: shared-head
 
-include::Sns-API.adoc[]
+include::api/Sns-API.adoc[]
 
-include::Member-API.adoc[]
+include::api/Member-API.adoc[]
 
-include::Local-API.adoc[]
+include::api/Local-API.adoc[]
 
-include::CampaignCategory-API.adoc[]
+include::api/CampaignCategory-API.adoc[]
 
-include::CampaignType-API.adoc[]
+include::api/CampaignType-API.adoc[]
 
-include::Campaign-API.adoc[]
+include::api/Campaign-API.adoc[]
 
-include::Application-API.adoc[]
+include::api/Application-API.adoc[]
 
-include::Applicant-API.adoc[]
+include::api/Applicant-API.adoc[]
 
-include::Review-API.adoc[]
-
-
-
-
-
+include::api/Review-API.adoc[]
