+++ src/test/java/com/bethefirst/lifeweb/util/RestDocsConfig.java
@@ -1,11 +1,10 @@
-package com.bethefirst.lifeweb;
+package com.bethefirst.lifeweb.util;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation;
 import org.springframework.restdocs.mockmvc.RestDocumentationResultHandler;
 import org.springframework.restdocs.operation.preprocess.Preprocessors;
-import org.springframework.restdocs.snippet.Attributes.Attribute;
 
 @TestConfiguration
 public class RestDocsConfig {
@@ -17,10 +16,6 @@ public class RestDocsConfig {
 				Preprocessors.preprocessRequest(Preprocessors.modifyUris().port(8088), Preprocessors.prettyPrint()),
 				Preprocessors.preprocessResponse(Preprocessors.prettyPrint())
 		);
-	}
-
-	protected static final Attribute field(final String key, final String value) {
-		return new Attribute(key,value);
 	}
 
 }
